requirements.txt
llama-index
pypdf
python-dotenv
ipykernel
torch
torchvision
pillow
python-dotenv
transformers
streamlit
streamlit-chat